feature: rename game use case instead of engine
