+++ app/src/main/java/com/amsterdam/aflami/di/viewModelModule.kt
@@ -2,7 +2,7 @@ package com.amsterdam.aflami.di
 
 import androidx.lifecycle.SavedStateHandle
 import com.amsterdam.viewmodel.cast.CastScreenArgs
-import com.amsterdam.viewmodel.game.GameArgs
+import com.amsterdam.viewmodel.game.whichGenre.GameArgs
 import com.amsterdam.viewmodel.listDetails.ListDetailsArgs
 import com.amsterdam.viewmodel.movieDetails.MovieDetailsArgs
 import com.amsterdam.viewmodel.seriesDetails.SeriesDetailsArgs
