+++ domain/src/main/java/com/amsterdam/domain/useCase/game/whichGenre/DoGuessGenreGameHintUseCase.kt
@@ -1,7 +1,11 @@
 package com.amsterdam.domain.useCase.game.whichGenre
 
-class DoGuessGenreGameHintUseCase {
-    operator fun invoke(movieReleasedDateQuestion: GetGenreQuestionsUseCase.MovieGenreQuestion): GetGenreQuestionsUseCase.MovieGenreQuestion {
+import com.amsterdam.domain.useCase.game.GetTotalUserPointsUseCase
+
+class DoGuessGenreGameHintUseCase(
+    private val getTotalUserPointsUseCase: GetUser
+) {
+    operator fun invoke(movieReleasedDateQuestion: GenerateMovieGenreQuestionsUseCase.MovieGenreQuestion): GenerateMovieGenreQuestionsUseCase.MovieGenreQuestion {
         val choices = movieReleasedDateQuestion.genreChoices.toMutableList()
 
         val wrongChoiceToRemove = choices
@@ -10,5 +14,9 @@ class DoGuessGenreGameHintUseCase {
 
         choices.remove(wrongChoiceToRemove)
         return movieReleasedDateQuestion.copy(genreChoices = choices)
+    }
+
+    companion object {
+        private const val REQUIRED_HINT_POINTS = 10
     }
 }
