+++ domain/src/main/java/com/amsterdam/domain/useCase/game/whichGenre/GuessMovieGenreUseCase.kt
@@ -4,21 +4,21 @@ import com.amsterdam.domain.timer.TimerHandler
 import com.amsterdam.domain.useCase.game.GetGameDifficultyByDifficultyTypeUseCase
 import com.amsterdam.entity.GameDifficulty
 
-class GuessMovieGenreGameEngin(
-    private val getGenreQuestionsUseCase: GetGenreQuestionsUseCase,
+class GuessMovieGenreUseCase(
+    private val generateMovieGenreQuestionsUseCase: GenerateMovieGenreQuestionsUseCase,
     private val getGameDifficultyByDifficultyTypeUseCase: GetGameDifficultyByDifficultyTypeUseCase,
     private val doGuessGenreGameHintUseCase: DoGuessGenreGameHintUseCase,
     private val timerHandler: TimerHandler
 ) {
-    val cachedQuestions: MutableList<GetGenreQuestionsUseCase.MovieGenreQuestion> = mutableListOf()
+    val cachedQuestions: MutableList<GenerateMovieGenreQuestionsUseCase.MovieGenreQuestion> = mutableListOf()
 
     suspend fun startGame(
         difficultyType: GameDifficulty.DifficultyType,
         onTimerUpdate: (remainingSeconds: Int) -> Unit,
         onTimeFinish: () -> Unit
-    ): List<GetGenreQuestionsUseCase.MovieGenreQuestion> {
+    ): List<GenerateMovieGenreQuestionsUseCase.MovieGenreQuestion> {
         val gameDifficulty = getGameDifficultyByDifficultyTypeUseCase(difficultyType)
-        val questions = getGenreQuestionsUseCase(
+        val questions = generateMovieGenreQuestionsUseCase(
             questionCount = gameDifficulty.totalQuestions
         )
         cachedQuestions.addAll(questions)
@@ -31,7 +31,7 @@ class GuessMovieGenreGameEngin(
         return questions
     }
 
-    fun useHint(questionId: Long): GetGenreQuestionsUseCase.MovieGenreQuestion {
+    fun useHint(questionId: Long): GenerateMovieGenreQuestionsUseCase.MovieGenreQuestion {
         val question = cachedQuestions.find { it.id == questionId }
         return doGuessGenreGameHintUseCase(question!!)
     }
