+++ viewModel/src/main/java/com/amsterdam/viewmodel/game/whichGenre/GameArgs.kt
@@ -1,14 +1,11 @@
-package com.amsterdam.viewmodel.game
+package com.amsterdam.viewmodel.game.whichGenre
 
 import androidx.lifecycle.SavedStateHandle
-import com.amsterdam.entity.Game
 import com.amsterdam.entity.GameDifficulty
 
 class GameArgs(
     savedStateHandle: SavedStateHandle,
 ) {
-    val gameType =
-        savedStateHandle.get<Game.GameType>("gameType") ?: Game.GameType.GUESS_MOVIE_BY_POSTER
     val difficulty = savedStateHandle.get<String>("difficulty")?.let { difficultyString ->
         try {
             GameDifficulty.DifficultyType.valueOf(difficultyString.uppercase())
