+++ viewModel/src/main/java/com/amsterdam/viewmodel/game/whichGenre/GameInteractionListener.kt
@@ -1,4 +1,4 @@
-package com.amsterdam.viewmodel.game
+package com.amsterdam.viewmodel.game.whichGenre
 
 interface GameInteractionListener {
     fun onCancelGameClick()
